Redesign the note filename creation process
Add `created_at` field to the Note dataclass. If title is empty, this
field is used to serve as a note and filename title. Otherwise the title
and timestamp are concacinated.

Regex sanitizing title is updated to remove illegal characters instead
replacing them with `_`, and each space is replaced with `_`, as
white-space in file names is annoying to handle.

More code in the whole file is updated to correctly handle having title
and timestamp as separate entities.